+++ .eslintrc
@@ -1,15 +1,20 @@
 {
-  "extends": [
-    "plugin:prettier/recommended",
-    "react-app",
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended"
-  ],
+  "parser": "@typescript-eslint/parser",
   "plugins": ["react", "@typescript-eslint"],
+  "env": {
+    "browser": true,
+    "node": true
+  },
+  "extends": ["eslint:recommended", "prettier", "plugin:react/recommended", "plugin:prettier/recommended"],
   "rules": {
-    "@typescript-eslint/interface-name-prefix": "on",
-    "@typescript-eslint/explicit-function-return-type": "on",
-    "@typescript-eslint/explicit-module-boundary-types": "on",
-    "@typescript-eslint/no-explicit-any": "on"
+    "@typescript-eslint/explicit-module-boundary-types": "off",
+    "prettier/prettier": ["error", { "endOfLine": "auto" }],
+    "react/react-in-jsx-scope": "off",
+    "no-unused-vars": "off"
+  },
+  "settings": {
+    "react": {
+      "version": "detect"
+    }
   }
 }
