+++ .prettierrc
@@ -1,9 +1,9 @@
 {
-    "singleQuote": true,
-    "trailingComma": "all",
-    "semi": true,
-    "useTabs": false,
-    "tabWidth": 2,
-    "printWidth": 80,
-    "arrowParens": "always"
+  "singleQuote": true,
+  "trailingComma": "all",
+  "semi": true,
+  "useTabs": false,
+  "tabWidth": 2,
+  "printWidth": 120,
+  "arrowParens": "always"
 }
