+++ src/api/pokeApi.ts
@@ -1,0 +1,6 @@
+import axios from 'axios';
+
+export const getPokeData = async () => {
+  const res = await axios.get(`https://pokeapi.co/api/v2/pokemon?limit=${LIMIT}&offset=0`);
+  return res.data;
+};
