+++ src/components/button/PageNationButton.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from "react";
+import { EventButton } from "./button.styled";
+
+interface ChildrenT {
+  children: ReactNode;
+}
+
+const PageNationButton = ({ children }: ChildrenT) => {
+  return (
+    <EventButton>{children}</EventButton>
+  );
+};
+
+export default PageNationButton;
