+++ src/components/button/button.styled.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components"
+
+export const EventButton = styled.button`
+  background-color: var(--button-bg-color);
+  border-radius: 10px;
+  padding: 10px 15px;
+  font-size: 1rem;
+  color: var(--button-text-color);
+`
